Reorganized tests as needed by vue-jest.
