+++ src/__tests__/ModifiedCard.spec.js
@@ -1,13 +1,13 @@
 import { mount } from '@vue/test-utils';
-import ModifiedIcon from './ModifiedIcon.vue';
+import ModifiedCard from '../components/cards/ModifiedCard.vue';
 
-describe('ModifiedIcon', () => {
+describe('ModifiedCard', () => {
   test('is initially invisible', () => {
-    const wrapper = mount(ModifiedIcon);
+    const wrapper = mount(ModifiedCard);
     expect(wrapper.findComponent("div").exists()).toBeFalsy();
   });
   test('remains invisible if the comparison values provided are identical', () => {
-    const wrapper = mount(ModifiedIcon, {
+    const wrapper = mount(ModifiedCard, {
       propsData: {
         compare: { key: 'test1' },
         compareTo: { key: 'test1' },
@@ -16,7 +16,7 @@ describe('ModifiedIcon', () => {
     expect(wrapper.findComponent("div").exists()).toBeFalsy();
   });
   test('becomes visible if the comparison values provided are different', () => {
-    const wrapper = mount(ModifiedIcon, {
+    const wrapper = mount(ModifiedCard, {
       propsData: {
         compare: { key: 'test1' },
         compareTo: { key: 'test2' },
@@ -24,8 +24,8 @@ describe('ModifiedIcon', () => {
     });
     expect(wrapper.findComponent("div").exists()).toBeTruthy();
   });
-  test('becomes visible if one of the comparison values is undefined', () => {
-    const wrapper = mount(ModifiedIcon, {
+  test('becomes visible if one of the comparison value is undefined', () => {
+    const wrapper = mount(ModifiedCard, {
       propsData: {
         compare: { key: 'test1' },
         compareTo: undefined,
